#919 refactor: extract shop search
